[onert] Refactor CPU backend to use operation list
This commit refactors the CPU backend to use operation list in CMakeLists.txt and KernelGenerator header
- Rename operation layer files to follow operation name
- Update CMakeLists.txt to explicitly list source files instead of using GLOB_RECURSE
- Add new implementation of QuantizeLayer
- Move MeanLayerand QuantizeLayer into ReduceLayer.h/cc and ElementwiseUnaryLayer.h/cc

ONE-DCO-1.0-Signed-off-by: Hyeongseok Oh <hseok82.oh@samsung.com>
