+++ runtime/onert/backend/cpu/KernelGenerator.cc
@@ -20,12 +20,12 @@
 #include "ops/ArgMinMaxLayer.h"
 #include "ops/BatchToSpaceNDLayer.h"
 #include "ops/BinaryArithmeticLayer.h"
-#include "ops/CompareLayer.h"
+#include "ops/ComparisonLayer.h"
 #include "ops/ConcatLayer.h"
-#include "ops/ConvolutionLayer.h"
+#include "ops/Conv2DLayer.h"
 #include "ops/DepthToSpaceLayer.h"
-#include "ops/DepthwiseConvolutionLayer.h"
-#include "ops/DynamicUpdateSlice.h"
+#include "ops/DepthwiseConv2DLayer.h"
+#include "ops/DynamicUpdateSliceLayer.h"
 #include "ops/ElementwiseActivationLayer.h"
 #include "ops/ElementwiseBinaryLayer.h"
 #include "ops/ElementwiseUnaryLayer.h"
@@ -34,15 +34,13 @@
 #include "ops/FullyConnectedLayer.h"
 #include "ops/GatherLayer.h"
 #include "ops/LSTMLayer.h"
-#include "ops/MeanLayer.h"
 #include "ops/DetectionPostProcessLayer.h"
 #include "ops/OneHotLayer.h"
 #include "ops/OperationUtils.h"
 #include "ops/PackLayer.h"
 #include "ops/PadLayer.h"
-#include "ops/PoolLayer.h"
+#include "ops/Pool2DLayer.h"
 #include "ops/PowLayer.h"
-#include "ops/QuantizeLayer.h"
 #include "ops/RangeLayer.h"
 #include "ops/RankLayer.h"
 #include "ops/ReduceLayer.h"
@@ -53,7 +51,7 @@
 #include "ops/SelectLayer.h"
 #include "ops/ShapeLayer.h"
 #include "ops/SliceLayer.h"
-#include "ops/SoftMaxLayer.h"
+#include "ops/SoftmaxLayer.h"
 #include "ops/StridedSliceLayer.h"
 #include "ops/SpaceToBatchNDLayer.h"
 #include "ops/SpaceToDepthLayer.h"
@@ -63,12 +61,12 @@
 #include "ops/TopKV2Layer.h"
 #include "ops/TransposeLayer.h"
 #include "ops/UnpackLayer.h"
-#include "ops/SquaredDiffLayer.h"
-#include "ops/L2NormLayer.h"
+#include "ops/SquaredDifferenceLayer.h"
+#include "ops/L2NormalizationLayer.h"
 #include "ops/BatchMatMulLayer.h"
 #include "ops/BroadcastToLayer.h"
 #include "ops/FusedBatchNormLayer.h"
-#include "ops/LogSoftMaxLayer.h"
+#include "ops/LogSoftmaxLayer.h"
 #include "ops/StatelessRandomUniformLayer.h"
 #include "ops/RmsNormLayer.h"
