+++ runtime/onert/backend/cpu/Operation.lst
@@ -1,0 +1,53 @@
+OP(AddN)
+OP(ArgMinMax)
+OP(BatchMatMul)
+OP(BatchToSpaceND)
+OP(BinaryArithmetic)
+OP(BroadcastTo)
+OP(Comparison)
+OP(Concat)
+OP(Conv2D)
+OP(DepthToSpace)
+OP(DepthwiseConv2D)
+OP(DynamicUpdateSlice)
+OP(ElementwiseActivation)
+OP(ElementwiseBinary)
+OP(ElementwiseUnary)
+OP(ExpandDims)
+OP(Fill)
+OP(FullyConnected)
+OP(FusedBatchNorm)
+OP(Gather)
+OP(L2Normalization)
+OP(LogSoftmax)
+OP(LSTM)
+OP(DetectionPostProcess)
+OP(OneHot)
+OP(Pack)
+OP(Pad)
+OP(Pool2D)
+OP(Pow)
+OP(Range)
+OP(Rank)
+OP(Reduce)
+OP(Reshape)
+OP(ResizeBilinear)
+OP(Reverse)
+OP(RmsNorm)
+OP(RoPE)
+OP(Select)
+OP(Shape)
+OP(Slice)
+OP(Softmax)
+OP(SpaceToBatchND)
+OP(SpaceToDepth)
+OP(Split)
+OP(SplitV)
+OP(SquaredDifference)
+OP(Squeeze)
+OP(StatelessRandomUniform)
+OP(StridedSlice)
+OP(TopKV2)
+OP(Tile)
+OP(Transpose)
+OP(Unpack)
