+++ runtime/onert/backend/cpu/ops/ComparisonLayer.cc
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "CompareLayer.h"
+#include "ComparisonLayer.h"
 
 #include "OperationUtils.h"
