+++ runtime/onert/backend/cpu/ops/L2NormalizationLayer.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2020 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -7,15 +7,15 @@
  *
  *      http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
+ * Unless required by applicable law or agreed to in riting, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
 
-#ifndef __ONERT_BACKEND_CPU_OPS_SOFTMAXLAYER_H__
-#define __ONERT_BACKEND_CPU_OPS_SOFTMAXLAYER_H__
+#ifndef __ONERT_BACKEND_CPU_OPS_L2NORMALIZATION_LAYER_H__
+#define __ONERT_BACKEND_CPU_OPS_L2NORMALIZATION_LAYER_H__
 
 #include <backend/IPortableTensor.h>
 
@@ -23,33 +23,24 @@
 
 namespace onert::backend::cpu::ops
 {
-
-class SoftMaxLayer : public ::onert::exec::IFunction
+class L2NormLayer : public ::onert::exec::IFunction
 {
 public:
-  SoftMaxLayer();
+  L2NormLayer() : _input(nullptr), _output(nullptr)
+  {
+    // Nothing
+  }
 
 public:
-  void softmaxFloat32();
-
-  template <typename T> void softmaxQuant8();
-
-  void configure(const IPortableTensor *input, const float beta, IPortableTensor *output);
+  void configure(const IPortableTensor *_input, IPortableTensor *output);
 
   void run() override;
 
-protected:
+private:
   const IPortableTensor *_input;
   IPortableTensor *_output;
-
-private:
-  float _beta;
-
-  float _table[256];
-  uint8_t _uint8_table1[256];
-  uint8_t _uint8_table2[256];
 };
 
 } // namespace onert::backend::cpu::ops
 
-#endif // __ONERT_BACKEND_CPU_OPS_SOFTMAXLAYER_H__
+#endif // __ONERT_BACKEND_CPU_OPS_L2NORMALIZATION_LAYER_H__
