+++ runtime/onert/backend/cpu/ops/ReduceLayer.h
@@ -68,6 +68,28 @@ private:
   ReduceType _reduceType;
 };
 
+class MeanLayer : public ::onert::exec::IFunction
+{
+public:
+  MeanLayer();
+
+public:
+  void MeanFloat32();
+
+  void MeanQuant8();
+
+  void configure(const IPortableTensor *input, const IPortableTensor *axes, IPortableTensor *output,
+                 bool keep_dims);
+
+  void run() override;
+
+protected:
+  const IPortableTensor *_input;
+  const IPortableTensor *_axes;
+  IPortableTensor *_output;
+  bool _keep_dims;
+};
+
 } // namespace onert::backend::cpu::ops
 
 #endif // __ONERT_BACKEND_CPU_OPS_REDUCESUMLAYER_H__
