+++ runtime/onert/backend/cpu/ops/SoftmaxLayer.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 Samsung Electronics Co., Ltd. All Rights Reserved
+ * Copyright (c) 2018 Samsung Electronics Co., Ltd. All Rights Reserved
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,35 +14,42 @@
  * limitations under the License.
  */
 
-#ifndef __ONERT_BACKEND_CPU_OPS_COMPARELAYER_H__
-#define __ONERT_BACKEND_CPU_OPS_COMPARELAYER_H__
+#ifndef __ONERT_BACKEND_CPU_OPS_SOFTMAX_LAYER_H__
+#define __ONERT_BACKEND_CPU_OPS_SOFTMAX_LAYER_H__
 
 #include <backend/IPortableTensor.h>
 
 #include <exec/IFunction.h>
-#include <ir/operation/Comparison.h>
 
 namespace onert::backend::cpu::ops
 {
 
-class CompareLayer : public ::onert::exec::IFunction
+class SoftMaxLayer : public ::onert::exec::IFunction
 {
 public:
-  CompareLayer();
+  SoftMaxLayer();
 
 public:
-  void configure(const IPortableTensor *lhs, const IPortableTensor *rhs,
-                 const ir::operation::Comparison::ComparisonType op_type, IPortableTensor *output);
+  void softmaxFloat32();
+
+  template <typename T> void softmaxQuant8();
+
+  void configure(const IPortableTensor *input, const float beta, IPortableTensor *output);
 
   void run() override;
 
-private:
-  const IPortableTensor *_lhs;
-  const IPortableTensor *_rhs;
+protected:
+  const IPortableTensor *_input;
   IPortableTensor *_output;
-  ir::operation::Comparison::ComparisonType _op_type;
+
+private:
+  float _beta;
+
+  float _table[256];
+  uint8_t _uint8_table1[256];
+  uint8_t _uint8_table2[256];
 };
 
 } // namespace onert::backend::cpu::ops
 
-#endif // __ONERT_BACKEND_CPU_OPS_COMPARELAYER_H__
+#endif // __ONERT_BACKEND_CPU_OPS_SOFTMAX_LAYER_H__
