+++ runtime/onert/backend/train/ops/MeanLayer.h
@@ -17,7 +17,7 @@
 #ifndef __ONERT_BACKEND_TRAIN_OPS_MEANLAYER_H__
 #define __ONERT_BACKEND_TRAIN_OPS_MEANLAYER_H__
 
-#include <ops/MeanLayer.h>
+#include <ops/ReduceLayer.h>
 #include <backend/IPortableTensor.h>
 
 #include <exec/train/ITrainableFunction.h>
